Update libtar with changes from cm-13.0
This fixes extractTarFork error 255 with large data backups &
other funny stuff going on such as having a linux chroot on your
data partition. I'm not sure which part of this commit fixes it,
but I'm glad I can back up my data partition now.

My best guess would be the fixes to long path names with symlinks
is what fixed the issue.

These changes also now allow me to restore ext4 backups to f2fs
partitions and vice-versa, and boot without any issues.

Added commits:

From 166d83a51e0c51abcea37694dbd7df92d03c1f56 Mon Sep 17 00:00:00 2001
From: Tom Marshall <tdm@cyngn.com>
Date: Wed, 2 Jul 2014 09:34:40 -0700
Subject: [PATCH] libtar: Fix hardlink extract

Do unlink for hardlink extract just as symlink does.  Failure to unlink
causes extract failures when the target exists.

From 8ec5627a8ff0a91724c6d5b344f0e887da922527 Mon Sep 17 00:00:00 2001
From: Tom Marshall <tdm@cyngn.com>
Date: Thu, 17 Apr 2014 09:39:25 -0700
Subject: [PATCH] libtar: Add methods for in-memory files

 * tar_append_file_contents: append header and buffer, similar to
   tar_append_file.

 * tar_append_buffer: append buffer only, similar to
   tar_append_regfile.

 * tar_extract_file_contents: extract regfile contents to buffer,
   similar to tar_extract_regfile.

From 6f1e8f6709505aaf3d7f04f9bb4eb9632b3f6786 Mon Sep 17 00:00:00 2001
From: Tom Marshall <tdm@cyngn.com>
Date: Tue, 2 Jun 2015 15:17:24 -0700
Subject: [PATCH] libtar: Support transparent compression

From e628c2025549a24018bc568351465130a05daafb Mon Sep 17 00:00:00 2001
From: Tom Marshall <tdm@cyngn.com>
Date: Thu, 11 Feb 2016 12:49:30 -0800
Subject: [PATCH] libtar: Make file sizes 64-bit clean

This lets us create and restore files up to 8gb (2^33-1).  Larger files
require an extension to the tar header's size field.

Note tar_append_file_contents and tar_append_buffer are still limited
to the platform definition of size_t to avoid breaking the ABI.

From e18b457ea1cbf6be1adc3b75450ed1c737cd82ea Mon Sep 17 00:00:00 2001
From: Tom Marshall <tdm@cyngn.com>
Date: Thu, 11 Feb 2016 16:24:40 -0800
Subject: [PATCH] libtar: Cleanup, secure, and extend numeric fields

From 9d133fd44abe462e4134d6a170e5871e56e081de Mon Sep 17 00:00:00 2001
From: Tom Marshall <tdm@cyngn.com>
Date: Tue, 16 Feb 2016 16:42:27 -0800
Subject: [PATCH] libtar: Store gzFile in a global to avoid non-portable
 casting

+++ libtar/basename.c
@@ -64,7 +64,7 @@ openbsd_basename(path)
 	while (startp > path && *(startp - 1) != '/')
 		startp--;
 
-	if (endp - startp + 1 > sizeof(bname)) {
+	if (endp - startp + 1 > (int)sizeof(bname)) {
 		errno = ENAMETOOLONG;
 		return(NULL);
 	}
